==> 1d_clean_india_survey.ipynb <==
# 1d. Cleaning SHRUG data from India

import numpy as np
import pandas as pd
import geopandas as gpd
import os
import json
import time
import matplotlib.pyplot as plt
import seaborn as sns

import shrug_processing
from importlib import reload
from shrug_processing import merge_raw_shrug_files, prepare_shrug_data_for_geom, split_urban_rural

import shrid_aggregation
from shrid_aggregation import loop_over_states

from multiprocessing import Pool

### This notebook processes shrug data into the shapefiles ready for downstream analysis.

shrug_stats.dta and shrid2.gpkg are the files that Paul Novosad shared directly with us via email. He said public version would be released (as of Sep 2022). 

# at what threshold to merge rural regions until -- we tried 14 (median area of rural regions) and 25
sq_km_thresh= 25

## 1. process the raw data and shapefiles into one dataframe

data_dir = '/data/mosaiks/replication/shrug/'

data_fp = os.path.join(data_dir,'shrug_stats.dta')
shapefile_fp = os.path.join(data_dir,'shrid2.gpkg')
merge_raw_shrug_files(data_fp, shapefile_fp, data_dir=data_dir)

# check the output csv 
shrug_csv_fp = f'{data_dir}/shrug.csv'
all_df = pd.read_csv(os.path.join(data_dir,'shrug.csv'))
all_df.head(2)

## 2. combine small rural shapefiles 

data_dir_condensed = data_dir+'shrug_condensed_geoms'
if not os.path.exists(data_dir_condensed):
    os.mkdir(data_dir_condensed)
    print(data_dir_condensed)
    

# load gdf and add projected area
gdf_by_shrid = prepare_shrug_data_for_geom(shrug_csv_fp)

# separate urban and rural gdfs
gdfs_split = split_urban_rural(gdf_by_shrid)
gdf_urban = gdfs_split['gdf_urban']
gdf_rural = gdfs_split['gdf_rural']

print(f'there are {len(gdf_urban)} urban shrids')
print(f'there are {len(gdf_rural)} rural shrids before merging')
print(f'median area of urban regions: {np.median(gdf_urban.proj_area) / 1e6} km^2')

print(f'median area of urban regions: {np.median(gdf_rural.proj_area) / 1e6} km^2')

chunked_file_dir = os.path.join(data_dir_condensed,f'thresh_{sq_km_thresh}_sq_km')
if not os.path.exists(chunked_file_dir):
    os.mkdir(chunked_file_dir)
    print('making dir: ',chunked_file_dir)

gdf_urban.to_file(os.path.join(data_dir_condensed,'shrug_urban.geojson'), driver='GeoJSON')

### 2.1 loop over states in parallel

# this is the loop that takes a while.
loop_over_states(gdf_rural, 
                 sq_km_thresh, 
                 chunked_file_dir, 
                 num_threads=20)

### 2.2  Read in the separate chunked files and compile

states_rural = np.sort(gdf_rural.state.unique())
# read in and compile

fps = [os.path.join(data_dir_condensed,'shrug_urban.geojson')] +\
      [os.path.join(chunked_file_dir,f'shrug_rural_state_{state}.geojson') for state in states_rural] 
      

dfs_by_state  = []
print(f'of {len(fps)}: ', end = '')
for i,fp in enumerate(fps):
    print(f'{i} ', end = '')
    df_by_state = gpd.read_file(fp)
    if fp.endswith('shrug_urban.geojson'):
        df_by_state['region'] = df_by_state['shrid'].apply(lambda x: x.split('-')[2])
    dfs_by_state.append(df_by_state.to_crs(epsg=4326))
print()

# concatenate results by state
shrug_condensed = pd.concat(dfs_by_state).reset_index(drop=True)
# put in the shrids_merged_str value for the urban shrids
shrug_condensed.loc[shrug_condensed.urban==True,'shrids_merged_str'] = shrug_condensed[shrug_condensed.urban].shrid

# perform some checks

# make sure all shrids are used by evaluating if there are the same number of unique shrids
num_ids_per_row = shrug_condensed.shrids_merged_str.apply(lambda x: len(x.split(',')))
print(f'same number of original shrids after merged? {(sum(num_ids_per_row) == len(gdf_by_shrid))}')

# assign rural to places with no urban
shrug_condensed.rename(columns={'rural':'has_rural',
                                'urban':'has_urban'},inplace=True)
shrug_condensed.loc[:,'rural'] = ~ shrug_condensed.loc[:,'has_urban'] 

# make new id and dicts explaining the relationships
shrug_condensed['condensed_shrug_id'] = np.arange(len(shrug_condensed))

# save output
shrug_condensed.to_file(os.path.join(data_dir,
                                     f'shrug_condensed_regions_{sq_km_thresh}.geojson'),driver='GeoJSON')
shrug_condensed.drop('geometry',axis=1).to_csv(os.path.join(data_dir,
                                     f'shrug_condensed_regions_{sq_km_thresh}.csv'))

replication_out_fn = '/data/mosaiks/replication/surveys/india/grouped.csv'
shrug_condensed.drop('geometry',axis=1).to_csv(replication_out_fn))

# if you want to just load the preloaded one:
#shrug_condensed = gpd.read_file(os.path.join(data_dir,
#                                f'shrug_condensed_regions_{sq_km_thresh}.geojson'))

print(f'there are {len(shrug_condensed)} total shrids after merging')
print(f'there are {sum(~shrug_condensed["rural"])} urban shrids after merging')
print(f'there are {sum(shrug_condensed["rural"])} rural shrids after merging')

### 2.3 Make and save dictionaries explaining the relationship between condensed shrids and original

new_id_to_shrids = {}
for key, val in zip(shrug_condensed['condensed_shrug_id'], 
                    shrug_condensed.shrids_merged_str.apply(lambda x: x.split(','))):
    new_id_to_shrids[key] = val

shrids_to_new_id = {}

for new_id, shrids_this_id in zip(shrug_condensed['condensed_shrug_id'], 
                    shrug_condensed.shrids_merged_str.apply(lambda x: x.split(','))):

    for shrid in shrids_this_id:
        shrids_to_new_id[shrid] = new_id
    
# save jsons
json.dump(shrids_to_new_id, open(os.path.join(data_dir,f'shrids_to_condensed_ids_{sq_km_thresh}.json'), 'w' ) )
json.dump(new_id_to_shrids, open(os.path.join(data_dir,f'condensed_ids_to_shrids_{sq_km_thresh}.json'), 'w' ) )
